refactor: IdPay prefix for configuration feature
